requirements.txt
opencv-python
numpy
Pillow
imagehash
yt-dlp
